RFI-2 Refactoring
