+++ OrariTreni/app/src/main/java/org/sglba/trainman/model/Station.java
