+++ OrariTreni/app/src/main/java/org/sglba/trainman/model/Vehicle.java
